xous-log: split platform-dependent code
