+++ services/xous-log/Cargo.toml
@@ -19,10 +19,17 @@ num-traits = {version = "0.2.14", default-features = false}
 
 utralib = {version = "0.1.15", optional = true, default-features = false }
 
+[target.'cfg(target_arch = "arm")'.dependencies]
+atsama5d27 = { git = "https://github.com/Foundation-Devices/atsama5d27.git", branch = "master", features = ["lcd-console"]}
+xous = { version = "0.9.33", features = ["v2p"] } # v2p feature is used when lcd-console feature is turned on
+
 [features]
 precursor = ["utralib/precursor"]
 hosted = ["utralib/hosted"]
 renode = ["utralib/renode"]
+atsama5d27 = ["utralib/atsama5d27"]
+
+lcd-console = []
 debugprint = [] # adding this allocates the UART for debugging the logger
 logging = [] # adding this allocates the hardware UART for console interactions
 #default = []
