+++ services/xous-log/src/platform/hosted/mod.rs
@@ -1,0 +1,1 @@
+pub mod implementation;
