+++ services/xous-log/src/platform/precursor/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "debugprint")]
+#[macro_use]
+pub mod debug;
+
+pub mod implementation;
